change window resolution to 1024 x 768

+++ include/render/BoardConstants.hpp
@@ -2,7 +2,7 @@
 
 class BoardConstants {
 public:
-    static constexpr int BOARD_SIZE = 8;
+    static constexpr uint8_t BOARD_SIZE = 8;
     static constexpr float SQUARE_SIZE = 75.0f;
     static const bool FLIPPED = false;
 };
